target/cortex_a: poll few times for target halted after reset
In case of secure boot, the DAP access is (eventually) opened a
little later after reset deassert. If OpenOCD is fast enough to
poll the target before this delay, it will issue the warning for
target "ran after reset and before halt ...".

Poll the target for few ms to catch the delay, before giving up.

Change-Id: Iaccf27262c3a5e56c9128261c66d1cc0925c0f6b
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>