target/cortex_a: reinit DP if not working anymore after reset
On some devices/boards the DP power-up flags CSYSPWRUPACK and/or
CDBGPWRUPACK get disabled at reset.
This would (partially) disable the DAP triggering errors later on.

Such issue could also be caused by a device/board that misses
	reset_config srst_pulls_trst
in the board configuration, but this case cannot be detected in
OpenOCD. Moreover, the option above is JTAG specific, and doesn't
have an equivalent one for SWD.

After reset check for DP access and for power-up flag status and,
in case needed, re-initialize the DP and inform the user about the
issue.

Change-Id: I1a43dc7933ed9a39e378b65a207ec49a748556b2
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>