drivers/stlink_usb: fix workaround for CSW
The data size 8/16/32-bits is encoded as value 1/2/4 in "size".
It worked fine for 32-bits just because the switch/case has the
"default" label on the right case for 32-bits.
In 8 and 16-bits mode it is crap!

Fix it and also convert unaligned 32-bits R/W in 8-bits; if the
address range crosses the auto-increment boundary then the driver
will split the R/W in a mix of 32 and 8-bits operations, that is
incompatible with this simple workaround.

Change-Id: Iab03b0397e53f9c76bc2a145f533ea1b0b655b1a
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>