target/cortex_a: fix R/W across non-contiguous virtual pages
When read_buffer/write_buffer uses AHB access, a virt2phys translation
is required to get the physical address for the AHB access.
But each translation is valid only within the limits of the page.
If the read_buffer/write_buffer crosses the page boundary, we cannot
assume that contiguous virtual pages are mapped on contiguous
physical pages, thus a new virt2phys translation is required at each
cross of page boundary.

Split the AHB access in pages and loop among them.

For simplicity, assume the page size is fixed to 4kB, that is the
minimum size allowed by armv7a. Actually, in this specific case
cortex_a_virt2phys() walks through the translation tables and can
provide the real page size. Add a comment to explain it for further
code improvement.

Change-Id: I1030fa35b3032bf23f2259c4ff820521e03fe154
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>