tcl/target/stm32mp15x_stpmic1: use non-secure access for backup register
Immediately after a reset the jtag port of STM32MP1 is open only in
non-secure mode, then reopened in secure mode. This prevents accessing
the AXI bus in secure mode for few milliseconds after reset.

The backup register is accessible in non-secure mode so, drop the
privileges before access it then restore them.

Change-Id: Ib3afccfcb133e1979faa59e4545328f193a0f969
Fixes: https://intbugzilla.st.com/show_bug.cgi?id=56373
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>