tcl/target/stm32mp15x: disable workarea
There is still not a common agreement on which memory area to use
as workarea, plus not clear how to use a static virtual address
during all the SW stages (TF-A, SPL, OP-TEE, U-Boot, Linux kernel
and Linux userland).
Today the workarea is not used, so let's disable it.
Also add the virtual address flag to remember we need this field
too.

Change-Id: I3585ecb095aad64a9ad0e0d0dc551a5d4fe2238d
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>