target/stm32mp15x: move the backup-flag handling in main script
The bit 16 of register tamp-backup[20] is reserved for passing
a flag, between debugger and application, that will be kept
across reset and power cycles. Today it is only used for PMIC
configuration, but can be used for other purposes.

Move the handling of the flag in the main script.
Rename the function to set/clear the flag with a more generic
name.

Change-Id: Ie778fa603d27dcdc052a67d0fbc46139d7b981cd
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>