target/stm32mp15x: use FSBL support for reset-halt
Some silicon revision of stm32mp15x require a mandatory power
cycle at reset. This makes vain any effort to deterministically
halt the cores after a reset.
Specific workaround has been implemented in FSBL (TF-A in case
of secure boot, U-Boot SPL in case of basic boot). This commit
instructs OpenOCD to cooperate with FSBL to halt the cores in
the workaround's code, thus simulating the halt in bootrom.

This commit can be removed if OpenOCD is used with a silicon
revision not affected by this limitation. Anyway, the script
does check the silicon revision and operates accordingly, so it
is safe to keep this commit whatever silicon revision is in use.

Change-Id: I8a5ac335e5238b4ac960ee1bd236839b94fb9a0a
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>