cortex_m: fix attach at target halted
The target could be left halted by a previous execution of the
debugger.
At a following execution of OpenOCD, the status of the target is
initialized to UNKNOWN and then set to HALTED in the poll event.
The transition UNKNOWN -> HALTED does not trigger the execution of
cortex_m_debug_entry() so the registers are not initialized.
If GDB attaches in this condition, running GDB command "info reg"
triggers an assert() in OpenOCD:
	openocd: ../src/target/armv7m.c:222: armv7m_read_core_reg:
	  Assertion `num < (int)armv7m->arm.core_cache->num_regs'
	  failed.
	Aborted (core dumped)

Fixed by calling cortex_m_debug_entry() when poll() detects the
transition UNKNOWN -> HALTED.
While there, rework the poll() function to remove some redundant
code and unnecessary parenthesis.

The hla_target does not have the same problem.

Change-Id: If6d9ec7c140b11bf3c3e76014db7e5f9f5998b03
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Fixes: https://sourceforge.net/p/openocd/tickets/216/