semihosting_fileio: fix null terminated string length for GDB
The semihosting code builds strings as null terminated, then pass
to the debugger the string pointer and the string length. The
length is computed through strlen(), thus doesn't include the
trailing zero character.
The trailing zero is properly taken in consideration by OpenOCD
for local file handling, but OpenOCD does not pass it correctly to
GDB in remote file I/O mode. In fact, as explicitly stated inside
a comment in GDB source code [1]:
	/* 1. Parameter: Ptr to pathname / length incl. trailing zero.  */
the length field should include the trailing zero character.

This issue affects the semihosting calls "open", "unlink",
"rename" and "system".

Fix it by adding a "+1" to the string length.

[1] http://sourceware.org/git/gitweb.cgi?p=binutils-gdb.git;f=gdb/remote-fileio.c;h=11c141e42c4d#l381

Change-Id: I35461bcb30f734fe2d51f7f0d418e3d04b4af506
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Fixes: c0e7ccbd87cf ("semihosting: support fileio operation")
Reviewed-on: https://gerrit.st.com/c/stm32ide/official/openocd/+/150121
Tested-by: Tarek BOUCHKATI <tarek.bouchkati@st.com>
Tested-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Tarek BOUCHKATI <tarek.bouchkati@st.com>