jtag: flush queue after reset for drivers using old reset model
Not all the jtag drivers have been migrated to the new reset model
and for those only we need to flush the jtag queue to make the
reset working with command 'adapter [de]assert ...'.

Add a queue flush and a FIXME comment to remove both when all the
drivers would be migrated.

Change-Id: Ib6667f987b1be2bce492841040302e742dd1cad1
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Reviewed-on: https://gerrit.st.com/c/stm32ide/official/openocd/+/160767
Tested-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@st.com>