cortex_a: don't wait for target halted in deassert_reset()
The tcl script src/target/startup.tcl has already the proper
centralized support to wait for targets to halt after a command
"reset halt". The extra wait in cortex_a_deassert_reset() is
not required.

Replace the indirect call to cortex_a_halt(), that embeds the wait
for halt, with a simple halt sequence.

Change-Id: Ic661c3791a29ba7d520e31f85a61f939a646feb5
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Reviewed-on: https://gerrit.st.com/c/stm32ide/official/openocd/+/160766
Tested-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@st.com>