target: armv8_dpm: silence error on register R/W
The command 'gdb_report_register_access_error' is used to silence
errors while reading registers and not reporting them to GDB.
Nevertheless, the error is printed by a LOG_ERROR() in armv8_dpm.

Change the message to LOG_DEBUG().
It will still cause the error to be propagated and eventually
printed by the caller (e.g. by the command 'reg').

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
Reviewed-by: Tarek BOCHKATI <tarek.bouchkati@st.com>
Tested-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
Tested-by: Tarek BOCHKATI <tarek.bouchkati@st.com>