target: add events read-buffer and write-buffer
The internal API target_read_buffer() and target_write_buffer()
are mainly used by the GDB connection for all the target's memory
read/write operations.

There are cases where the user needs to override these API, e.g.:
- on emulated targets, where it's too slow accessing the memory
  through the CPU and some specific workaround for direct access
  to the emulated memory could be provided by the emulator [1];
- to access HW registers while the target cannot provide such
  access, e.g. on not halted ARM Cortex-A or Cortex-R.

By abusing of the target events, add two new events 'read-buffer'
and 'write-buffer' that will be called to override the API above.
Since we cannot pass arguments to the body of the events, trick it
by requiring to have, as body of these events, the name of a Tcl
proc; the caller of the event will add the arguments.

Add two helper commands:
- $target_name read_buffer address count
- $target_name write_buffer address count data
that generate a new call to the internal API. These are useful to
forward the API to another target (e.g. mem_ap) or to modify the
arguments.

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Link: [1] https://review.openocd.org/c/openocd/+/7149
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
Tested-by: Tarek BOCHKATI <tarek.bouchkati@st.com>
Reviewed-by: Tarek BOCHKATI <tarek.bouchkati@st.com>
Tested-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>